requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
